// src/interfaces/proyectos/featureSection.ts
export interface FeatureSection {
  title: string;
  generalDescription: string;
  modelsDescription: string;
  kitchenDescription: string;
  luxuryDescription: string;
  amenitiesDescription: string;
  floorplansDescription: string;
  terrainDescription: string;
  tipoPropiedad: string[];
  materialConstruction: string;
}